Stop reading input file when all required data is found.
Pagemeter and swapmeter unnecessarily read to the end of the input file.

+++ linux/pagemeter.h
@@ -19,7 +19,7 @@ public:
 
   void checkResources( void );
 protected:
-  float pageinfo_[2][2];
+  unsigned long pageinfo_[2][2];
   int pageindex_;
   float maxspeed_;
   bool _vmstat;
